Upgraded emacs packages, and removed obsolete packages and configs.
Remove `cl-lib-highlight`. `cl-lib-highlight` is a package to
highlight `cl-lib` macros, and is obsolete since Emacs 24.4.

Remove `org-fs-tree` because it is not maintained now, and it is not
used.

Remove `hl-sexp` because it is obsolete.

Replace `gitignore-mode` and `gitconfig-mode` with `git-modes`,
because the latter replaces the former.

Replace `highlight-symbol` with `auto-highlight-symbol` because the
latter is in melpa.